Fix for app execution on variant boards
This provides a temporary solution for the problem where the
application does not work on the board after the upload. This is
true for boards that have variant source files (override of a
generic platform). Issue is due to the fact that the linker
dropped the object files of the variant sources as it already
found the corresponding weak symbols in the core library. The fix
for this requires linking the variant objects explicitly rather than
as part of core library. This commit fixes this temporarily using
the object library feature of CMake in an alternative way. It probably
works in CMake 3.1 and above, but requires invoking cmake second time,
and then doing build again, after the first successful build.