Improved library search algorithm (#22)
* Improved library search algorithm

Library and include names in some of the libraries differ, which
prevented the toolchain in locating the library (issue #19). The
improved algorithm introduces library indexer to index the available
libraries, their names and includes, the library search algorithm
utilizes this info during the search to fix the issue #19.

Also the internal target names of Arduino libraries are made consistent
as there is now a possibility to use both the include as well as library
names in target_link_arduino_libraries. This is to avoid multiple
targets of the same library.