Add basic support for arduino-cli detection
This adds the most basic of support for the environment created from
downloading the Arduino SDK via the official [arduino-cli project][1].

It fixes two issues encountered with detection:

1. If `set(ARDUINO_INSTALL_PATH "$env{LOCALAPPDATA}/Arduino15` is used,
   it will work, however the make system will loudly complain that it is
   unable to read `lib/version.txt`, however generation will succeed.
2. By default, the cli installation is not detected, so this adds the
   install path hint of the `$env{LOCALAPPDATA}/Arduino15` path and adds
   basic detection of `arduino-cli.yaml` which is created from the cli
   setup process.

The Linux update has been validated using Debian Buser running under
WSL1 and all tools work as far as I can tell, including detection,
build, sizing, and upload.

I do not have a Mac available to test, but I am basing the assumption of
`$ENV{HOME}/Library/Arduino15` as the install point since that's in the
package path as well.

Using the tool, running a command like `arduino-cli core install arduino:samd`
will download the tooling, the core, and set up the environment.

Originally I was detecting `arduino-cli.yaml` as the marker file but
under a brand new installation on Linux that file did not exist, but
`inventory.yaml` did.

I was unable to detect any version information of a matching SDK or
arduino-cli version, so the assumption of the file existing was enough
to trigger the new flow.

[1]: https://github.com/arduino/arduino-cli