Fix registry bug when globbing for sketchbook
The existing code fails on Windows installations as it was checking
`HKEY_LOCAL_MACHINE` instead of the user defined/customized
`HKEY_CURRENT_USER` as the `Personal` registry value only exists in the
user registry.

Additionally, the `file` command does not seem to resolve the registry
directive, so an additional call to `get_filename_component` was added
to resolve the directory path.