Add support for cli-installed libs without IDE
If a library is installed from the Arduino CLI via `arduino-cli lib
install <lib>` on Windows, the default location for the installed
library is under `Documents/Arduino/Libraries/<lib>`. When the IDE is
not installed, a preferences file pointing to this location is not
created and is not recognized by the build system.

This attempts to resolve the issue by detecting a missing sketchbook
location and the existance of the `_reg_documents` variable (resolved
when setting up package paths) by resolving the location and setting the
appropriate `ARDUINO_SKETCHBOOK_PATH` variable, allowing library
searches to find libraries installed by the CLI.